requirements.txt
fastapi==0.77.1
uvicorn==0.17.6
requests==2.27.1
pydantic==1.9.1
PyYAML==6.0